+++ src/main/resources/db/migration/V1_4__table_main_blocks.sql
@@ -1,4 +1,4 @@
 CREATE TABLE main_blocks (
-  id        BIGINT PRIMARY KEY,
+  id        INTEGER PRIMARY KEY,
   signature VARCHAR NOT NULL
 );
